Rollover from other branch.  Summary of changes below:
These changes are meant to allow you to use the UI Draggers to permanently adjust characters in an hscene to a more appropriate location when placed incorrectly due to using a non-default player height or for other reasons.  Once everything is adjusted, these offsets can be saved and then automatically applied the next time that hscene position is loaded with those characters.

UI Draggers now moves characters relative to their coordinate system and not the world coordinate system.  This was a necessary change for saving offsets, but I also think it makes adjusting characters easier because their X/Z movements are more predictable.
Added a Reset button to UI.  This button will reset the selected characters to the home (0) position.
Added a Save button to the UI.  This button will save the current offsets for all characters in the hscene for the current position and current motion of that hscene.  Offsets are saved on a position/motion/character pair basis for all characters in the hsene.
Added a Defaultbutton to the UI.  This button will save the current offsets for all characters in the hscene for the current position of that hscene.  When offsets are loaded, it will first look for the correct position/motion.  If it cannot find the motion, then it will use this "default" value.  Useful for making an initial adjustment to all motions in a position, before fine tuning individual motions.

Added a checkbox to the plugin setup to Enable/Disable auto loading of offsets.  If enabled, when entering a new hscene position, it will automatically search all saved offsets and apply them if a saved one is found.

Notes:
Characters are saved by their full name in game, so having multiple characteers with the same full name will cause issues.  Character order matters, so Character 1 as male and Character 2 as female is not the same as Character 2 as male and Character 1 as female, those would be separate character pairs and have their own unique offsets.
Position refers to the position of the hscene (ex. Missionary, Cowgirl, etc.)
Motion refers to the loop running for that position (ex. Idle, Wloop, Climax, etc).
The offsets file is located in the game root directory as AI_BetterHScenes.xml.  It is recommended not to edit this unless you know what you are doing.